Fix errors in Typescript
