+++ src/elements/index.js
@@ -4,7 +4,7 @@ export { default as Button } from './Button';
 export { default as Input } from './Input';
 export { default as TextArea } from './TextArea';
 export { default as Switch } from './Switch';
-export { default as Notification } from './Notification';
+export { default as Notification } from './Notification/Notification';
 export { default as Typography } from './Typography';
 export { default as DatePicker } from './DatePicker';
 export { default as DayPicker } from './DayPicker';
